connectors/bright_data/requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.3.0
boto3==1.38.27
botocore==1.38.27
branca==0.8.2
brightdata-sdk==1.1.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
duckdb==1.4.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
fivetran_connector_sdk==2.3.1
folium==0.20.0
fonttools==4.60.1
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
multidict==6.6.3
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.7.1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pipreqs-fivetran==1.0.1
platformdirs==4.5.0
plotly==6.3.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pymongo-migrate==1.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.3
rpds-py==0.28.0
s3transfer==0.13.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
uqlm==0.1.3
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xyzservices==2025.10.0
yarg==0.1.10
yarl==1.20.1
